Initial lookup and getId api's for miniId
