+++ miniserver/app.js
@@ -8,6 +8,7 @@ const cookieParser = require('cookie-parser')
 const logger = require('morgan')
 const config = require('./config')
 const _index = require('./routes/index')
+const _minIID = require('./routes/miniID')
 const bodyParser = require('body-parser')
 const app = express()
 const https = require('https')
@@ -96,6 +97,7 @@ if (config.corsOrigins) {
 app.use(express.static(path.join(__dirname, 'public')))
 app.options('*', async (_req, res) => res.end())
 app.use('/', _index)
+app.use('/id', _minIID)
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
