+++ miniserver/routes/miniID.js
@@ -1,21 +1,33 @@
 const express = require('express')
 const { StatusCodes } = require('http-status-codes')
+// const { NFT } = require('../../server/src/data/nft')
 const router = express.Router()
-const { hasUserSignedBody, validateID } = require('../../server/routes/middleware')
+// const { hasUserSignedBody } = require('../../server/routes/middleware')
+// const sharedUtils = require('../../shared/utils')
+const { validateID } = require('./middleware')
 const blockchain = require('../blockchain')
+
+// allows an existing user to lookup another user's address by their phone number, iff the phone number exists and the target user does not choose to hide its phone-address mapping (under `hide` parameter in settings)
+router.post('/lookup', validateID, async (req, res) => {
+  console.log(`req.body: ${JSON.stringify(req.body)}`)
+  const { phone, address } = req.processedBody
+  return res.json({ phone: phone, address: address })
+})
 
 router.post('/getid', validateID, async (req, res) => {
   // get MiniId contract
   const { phone, address } = req.processedBody
-  let validMiniID
+  const validMiniID = {}
   const miniID = blockchain.getMiniID()
+  //   const tokenCount = await miniID.balanceOf('0x90F79bf6EB2c4f870365E785982E1f101E93b906')
   const tokenCount = await miniID.balanceOf(address)
   if (tokenCount < 1) {
-    return res.status(StatusCodes.BAD_REQUEST).json({ error: `${phone} address(${address}) has no MiniId` })
+    // return res.status(StatusCodes.BAD_REQUEST).json({ error: `${phone} address(${address}) has no MiniId` })
+    return res.json({ success: false, error: `${phone} address(${address}) has no MiniId`, miniID: validMiniID })
   }
-  validMiniID.address = miniID.address()
-  validMiniID.tokenID = await miniID.tokenOfOwnerByIndex(address, 0)
-  validMiniID.tokenURI = await miniID.tokenOfOwnerByIndex(validMiniID.tokenID)
+  validMiniID.address = await miniID.address
+  validMiniID.tokenID = (await miniID.tokenOfOwnerByIndex(address, 0)).toNumber()
+  validMiniID.tokenURI = await miniID.tokenURI(validMiniID.tokenID)
 
   return res.json({ success: true, miniID: validMiniID })
 })
@@ -26,7 +38,7 @@ router.post('/getid', validateID, async (req, res) => {
 // signature (optional): a signature from the user
 // application (optional) : application to add tracking for
 // Returns
-// nftID: token address, tokenId, tokenqty (Optional)
+// nftID: tokenAddress, tokenId, tokenQty (Optional)
 // Logic Overview
 // Validate the id is a valid SMS id
 // Validate the signature is from the address of the ID
@@ -41,8 +53,14 @@ router.post('/getid', validateID, async (req, res) => {
 // Operator Mints the token (if this fails retrieve latest token and try again)
 // Update SMS Tracking Info
 // Return nftId
-router.post('/mintID', hasUserSignedBody, async (req, res) => {
-
+// router.post('/mintID', validateID, hasUserSignedBody, async (req, res) => {
+router.post('/mintID', validateID, async (req, res) => {
+  console.log(`req.body: ${JSON.stringify(req.body)}`)
+  const { phone, address } = req.processedBody
+  let tokenAddress
+  let tokenID
+  let tokenQty
+  return res.json({ phone: phone, address: address, tokenAddress, tokenID, tokenQty })
 })
 
 module.exports = router
