+++ miniwallet/deployments/ethLocal/Mini1155.json
@@ -1367,7 +1367,7 @@
     "transactionIndex": 0,
     "gasUsed": "3855279",
     "logsBloom": "0x00000000100000000000000000000000000000000040000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000020000000000000100000800000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000002000000000000000000020000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0xf8062d2f61c42bd95d04141d85691a061e4b8c1af7021b754fb7c46896545c4e",
+    "blockHash": "0xbb5633d65345bd63aa115851866d26d5644b9a420afe5feee75fafcbae045221",
     "transactionHash": "0xe2ed76eb84b3112d7ebc690363741928c0ae561af4cc0b64f6a4e8b54c4c742e",
     "logs": [
       {
@@ -1382,7 +1382,7 @@
         ],
         "data": "0x",
         "logIndex": 0,
-        "blockHash": "0xf8062d2f61c42bd95d04141d85691a061e4b8c1af7021b754fb7c46896545c4e"
+        "blockHash": "0xbb5633d65345bd63aa115851866d26d5644b9a420afe5feee75fafcbae045221"
       }
     ],
     "blockNumber": 12,
