+++ miniwallet/deployments/ethLocal/Mini721.json
@@ -1293,7 +1293,7 @@
     "transactionIndex": 0,
     "gasUsed": "3323050",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000800000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000010000000010000000000000000000000020000000000000100000800000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000002000000000000000000020000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0x696ca19176ab74bde876bcaa04f60e6686782297e53b4c14e1d0964f21ac2994",
+    "blockHash": "0xf2f4044bce0282622a52c2e15fb1be3196d6b06f766572cfab84d39b90e4aaf4",
     "transactionHash": "0xd2c86e9fa0dbc88d9a04e59d4c4bb465d8499d7a968a5108baf12fc826a4792b",
     "logs": [
       {
@@ -1308,7 +1308,7 @@
         ],
         "data": "0x",
         "logIndex": 0,
-        "blockHash": "0x696ca19176ab74bde876bcaa04f60e6686782297e53b4c14e1d0964f21ac2994"
+        "blockHash": "0xf2f4044bce0282622a52c2e15fb1be3196d6b06f766572cfab84d39b90e4aaf4"
       }
     ],
     "blockNumber": 8,
