+++ miniwallet/deployments/ethLocal/MiniID_Implementation.json
@@ -1032,7 +1032,7 @@
     "transactionIndex": 0,
     "gasUsed": "3586485",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000",
-    "blockHash": "0x12f7fef58574aa8c2c161a0468c8fdc815ad0f65385bde323bfb30c0a4f1aa47",
+    "blockHash": "0x2c241371ffb4ee14702966462fa070c5a951441b99a7ef97ca7d5c3f1f04d176",
     "transactionHash": "0x38891ae9ac44d4355fa5a32d6832fd4a89fcc8546fa33a6197c45b0e75a13bfb",
     "logs": [
       {
@@ -1045,7 +1045,7 @@
         ],
         "data": "0x00000000000000000000000000000000000000000000000000000000000000ff",
         "logIndex": 0,
-        "blockHash": "0x12f7fef58574aa8c2c161a0468c8fdc815ad0f65385bde323bfb30c0a4f1aa47"
+        "blockHash": "0x2c241371ffb4ee14702966462fa070c5a951441b99a7ef97ca7d5c3f1f04d176"
       }
     ],
     "blockNumber": 3,
