+++ miniwallet/deployments/ethLocal/MiniID_Proxy.json
@@ -79,7 +79,7 @@
     "transactionIndex": 0,
     "gasUsed": "468434",
     "logsBloom": "0x00000004000000000000000000000000400000000000000000000000000000000000000000000000000400008000084000400000000000000000000000000001000000000000000000000000000002000000000000000000000000000000400000000000020000000000000100000800000000000000000000010000001000000000000008000000000000000000008000000000000080000000000000000000000000000000000000000000000400000000000000000000001000000000000000000020000000200000000000040000001000002000000100002000000020000080000000000000000000000000000000000400000000000000000000000000",
-    "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff",
+    "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008",
     "transactionHash": "0x91ad5d8710ff9c262b0044b651bbb7aafa36705017ebce13f32352fe7d32fe58",
     "logs": [
       {
@@ -93,7 +93,7 @@
         ],
         "data": "0x",
         "logIndex": 0,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       },
       {
         "transactionIndex": 0,
@@ -108,7 +108,7 @@
         ],
         "data": "0x",
         "logIndex": 1,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       },
       {
         "transactionIndex": 0,
@@ -123,7 +123,7 @@
         ],
         "data": "0x",
         "logIndex": 2,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       },
       {
         "transactionIndex": 0,
@@ -138,7 +138,7 @@
         ],
         "data": "0x",
         "logIndex": 3,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       },
       {
         "transactionIndex": 0,
@@ -153,7 +153,7 @@
         ],
         "data": "0x",
         "logIndex": 4,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       },
       {
         "transactionIndex": 0,
@@ -165,7 +165,7 @@
         ],
         "data": "0x0000000000000000000000000000000000000000000000000000000000000001",
         "logIndex": 5,
-        "blockHash": "0x5f78aa53f4464e38ee9cd00ffeb030fbff130fd9b6c9a10d8657b66332c553ff"
+        "blockHash": "0x97cfc75cde62728e4cb61d43b3f9df08cf63dd447a25d215f2f59508a36c0008"
       }
     ],
     "blockNumber": 4,
