+++ miniwallet/deployments/ethLocal/MiniWallet_Implementation.json
@@ -952,7 +952,7 @@
     "transactionIndex": 0,
     "gasUsed": "2071284",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0x94536d11baa514b754a36ef2f08353078fd2313094def272b03cd7830628dc3e",
+    "blockHash": "0x68910b4753b4fe7bfeae10f8cd5de15956ac1f220f5f6dcedd9da50b5dbe40cb",
     "transactionHash": "0xd44c9c54ccc53dd82c064381048acd7f17ee179b3b8c5bfc8520b91cbbb5bd80",
     "logs": [],
     "blockNumber": 1,
