+++ miniwallet/deployments/ethLocal/MiniWallet_Proxy.json
@@ -153,7 +153,7 @@
     "transactionIndex": 0,
     "gasUsed": "1066226",
     "logsBloom": "0x00000004000000000002000000000000000000000000000000800000000000000200000000000000000400000000000010000000000000000000000000100000000000000000000000000000000000000003000000000000000000000000000000000800020000000000000100800800000804000000000000000000000000400000000000000000000000000000000000000000000080000400000000000000000000000000200000000000000600000000000000008000001000000000000000000000000000200000000000041000400000002008000100000000000020000000000000000000000410000000000001000000000000800000000000000000",
-    "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de",
+    "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c",
     "transactionHash": "0x65b5235684423a201571661f372e5023d8f85e0894c1d37e4809aefe904fade8",
     "logs": [
       {
@@ -168,7 +168,7 @@
         ],
         "data": "0x",
         "logIndex": 0,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -183,7 +183,7 @@
         ],
         "data": "0x",
         "logIndex": 1,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -198,7 +198,7 @@
         ],
         "data": "0x",
         "logIndex": 2,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -213,7 +213,7 @@
         ],
         "data": "0x",
         "logIndex": 3,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -228,7 +228,7 @@
         ],
         "data": "0x",
         "logIndex": 4,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -240,7 +240,7 @@
         ],
         "data": "0x0000000000000000000000000000000000000000000000000000000000000001",
         "logIndex": 5,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       },
       {
         "transactionIndex": 0,
@@ -254,7 +254,7 @@
         ],
         "data": "0x",
         "logIndex": 6,
-        "blockHash": "0x13240660fa4833582772963f1d2031e2b47fa0ba8654a81f6d322948e0ded2de"
+        "blockHash": "0x1cf2eed9d34dfb56f9f5ff78d7d3fbc9f638a7aded2de6a2504c9727a6a64f6c"
       }
     ],
     "blockNumber": 2,
